+++ test/validate.test.js
@@ -15,6 +15,14 @@ for (var name in geojsonFixtures) {
 test('roundtrip custom properties', roundtripTest(getJSON('props.json')));
 test('roundtrip single-ring MultiPolygon', roundtripTest(getJSON('single-multipoly.json')));
 
+test('roundtrip valid closed polygon with high-precision coordinates', function (t) {
+    var geojson = getJSON('precision.json');
+    var pbf = new Pbf(geobuf.encode(geojson, new Pbf()));
+    var ring = geobuf.decode(pbf).features[0].geometry.coordinates[0];
+    t.same(ring[0], ring[4]);
+    t.end();
+});
+
 function roundtripTest(geojson) {
     return function (t) {
         t.same(geobuf.decode(new Pbf(geobuf.encode(geojson, new Pbf()))), geojson);